Added a lot of code to support blocking/restarting on a by-Behaviour basis. Now Behaviour is an abstract class, providing foundations for a bidirectional Chain Of Responsibility based notification system. Each behaviour has a boolean state variable telling whether it is runnable or not.
A protected inner class is used to represent events that are sent when
a Behaviour changes its runnable state; this event can be sent both
upwards and downwards in behaviour aggregation hierarchy. The
notification system is totally transparent and automagical; a pair of
public methods is provided to perform state transitions: these are
called block() and restart().